moved up section variables and columns filter
